+++ pywren/pywren_ibm_cloud/storage/cleaner.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-from pywren_ibm_cloud.storage import storage_internal
+from pywren_ibm_cloud.storage import storage
 import logging
 import sys
 import os
@@ -28,7 +28,7 @@ def clean_bucket(bucket, prefix, storage_config):
     Wrapper of clean_os_bucket(). Use this method only when storage_config is
     in JSON format. In any other case, call directly clean_os_bucket() method.
     """
-    internal_storage = storage_internal.Storage(json.loads(storage_config))
+    internal_storage = storage.InternalStorage(json.loads(storage_config))
     sys.stdout = open(os.devnull, 'w')
     clean_os_bucket(bucket, prefix, internal_storage)
     sys.stdout = sys.__stdout__
