+++ pywren/pywren_ibm_cloud/wrenutil.py
@@ -192,7 +192,7 @@ def split_path(path):
 
 def verify_args(func, data, object_processing=False):
     # Verify parameters
-    none_verify_parameters = ['storage', 'internal_storage']
+    none_verify_parameters = ['storage', 'ibm_cos', 'swift', 'internal_storage']
     func_sig = inspect.signature(func)
     
     # Check mandatory parameters in function
