+++ runtime/deploy_runtime
@@ -3,7 +3,7 @@ import click
 import os
 import json
 from pywren_ibm_cloud import wrenconfig
-from pywren_ibm_cloud.storage import storage_internal
+from pywren_ibm_cloud.storage import storage
 from pywren_ibm_cloud.cf_connector import CloudFunctions
 from pywren_ibm_cloud.wrenconfig import CF_ACTION_NAME_DEFAULT
 
@@ -51,7 +51,7 @@ def create_runtime(image_name):
 
     # Create storage_handler to upload modules file
     storage_config = wrenconfig.extract_storage_config(config)
-    internal_storage = storage_internal.Storage(storage_config)
+    internal_storage = storage.InternalStorage(storage_config)
     internal_storage.put_runtime_info(runtime_name, runtime_meta)
 
     # Create zipped PyWren action
